backend/requirements/install.txt
# FitTracker Backend - Quick Install Requirements
# Este archivo contiene solo las dependencias esenciales para que funcione la aplicación

# Core Django
Django==5.0.2
djangorestframework==3.15.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Environment variables
python-decouple==3.8

# CORS for frontend integration
django-cors-headers==4.3.1

# Filtering and searching
django-filter==24.1

# API Documentation
drf-spectacular==0.27.1

# HTTP requests (CRÍTICO - faltaba en el requirements original)
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Timezone handling
pytz==2023.3

# Database (opcional - solo si usas MySQL)
# mysqlclient==2.2.4

# Para desarrollo básico
django-extensions==3.2.3

# Image processing básico
Pillow==10.2.0





